Manual backup channel change for Bulk-Edit APs
